No real changes in makefile but avoid repetitious in bakefile.
Collect all preprocessor options needed for Scintilla compilation in
wxscintilla_cppflags template and use for both Scintilla compilation itself
and for the wx libraries using it.

The order of flags in the makefiles has changed but they are still the same
except for Borland makefile which now includes Borland-specific warning
suppression option everywhere it's needed and not just in multilib targets.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@65089 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775